showPortal refactor
